+++ src/main/java/com/marginallyclever/ro3/apps/viewport/OpenGLPanel.java
@@ -7,7 +7,7 @@ import com.marginallyclever.convenience.helpers.MatrixHelper;
 import com.marginallyclever.convenience.helpers.ResourceHelper;
 import com.marginallyclever.ro3.Registry;
 import com.marginallyclever.ro3.apps.viewport.renderpass.RenderPass;
-import com.marginallyclever.ro3.apps.viewport.viewporttools.ViewportTool;
+import com.marginallyclever.ro3.apps.viewport.viewporttool.ViewportTool;
 import com.marginallyclever.ro3.node.nodes.pose.poses.Camera;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
